use a static_cast from signed->unsigned integer in LuminaX11.cpp to silence a possible c++11 build warning/error.

+++ src-qt5/core/libLumina/LuminaX11.cpp
@@ -778,7 +778,7 @@ void LXCB::MoveResizeWindow(WId win, QRect geom){
 // ===== ResizeWindow() =====
 void LXCB::ResizeWindow(WId win, int width, int height){
   //Use the basic XCB functions instead of ewmh
-  uint32_t values[] = {width, height};
+  uint32_t values[] = { static_cast<unsigned int>(width), static_cast<unsigned int>(height) };
   xcb_configure_window(QX11Info::connection(), win, XCB_CONFIG_WINDOW_WIDTH | XCB_CONFIG_WINDOW_HEIGHT, values);
 }
